+++ frontends/web/src/routes/market/components/markettab.tsx
@@ -1,26 +1,43 @@
 // SPDX-License-Identifier: Apache-2.0
 
+import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { PillButton, PillButtonGroup } from '../../../components/pillbuttongroup/pillbuttongroup';
+import type { AccountCode, TAccount } from '@/api/account';
 import type { TMarketAction } from '@/api/market';
+import { useMarketContext } from '@/routes/market/market-context';
+import { PillButton, PillButtonGroup } from '@/components/pillbuttongroup/pillbuttongroup';
 import { NewBadge } from '@/components/new-badge/new-badge';
 import style from './markettab.module.css';
 
 export type TMarketplaceTab = TMarketAction | 'insure';
 
 type TProps = {
-  onChangeTab: (tab: TMarketplaceTab) => void;
+  accounts: TAccount[];
   activeTab: TMarketplaceTab;
-  showSwap: boolean;
+  code: AccountCode;
 };
 
-
 export const MarketTab = ({
-  onChangeTab,
+  accounts,
   activeTab,
-  showSwap,
+  code,
 }: TProps) => {
   const { t } = useTranslation();
+  const navigate = useNavigate();
+  const { showSwap } = useMarketContext();
+
+  const onChangeTab = (tab: TMarketplaceTab) => {
+    if (tab === 'insure') {
+      navigate(
+        accounts.some(({ bitsuranceStatus }) => bitsuranceStatus)
+          ? `/market/bitsurance/dashboard/${code}`
+          : `/market/bitsurance/${code}`
+      );
+      return;
+    }
+    navigate(`/market/select/${code}?tab=${tab}`);
+  };
+
   return (
     <PillButtonGroup className={style.navigation} size="large">
       <PillButton
